Restructuring + Various updates
Restructuring the application due to issues with using "pages" and "app" folders at the same time. This is not good practice and was giving me more and more issues so was about time I repaired this. Fixed the navbar so it actually works across pages, moved userbutton into this navbar. Various other minor tweaks not worth mentioning here