Adding journey logging into firebase
You can now log your journey information into firebase! This will be developed further in a future update but for now it's stored in the database. Various other edits were also done such as turning the form to add your journey into a component. The page showing the emissions and some data involved with them was updated to have a button to toggle between Co2 being shown or not as it takes up a stupidly large amount of the graph.