Updated AddData and getData
Added a page reload to the addData page so it refreshes the information automatically. Added a fuel type to each part of the getData page so you know what journey was completed with which fuel type. And finally, updated the AQI page to have a key and an explanation about what the levels mean and what AQI means