Many updates
Updates the aqi page to dynamically change colour to help indicate what aqi level the city you input is currently in. Updated the 3d model page to show proper advantages to combustion engines. Removed the pointless HC emissions that weren't provided for the emissions page. Changed colours to be less horrible on this page too. Removed some details from the footer as some testers who decided to spam email me made clear that this was a terrible idea. Vastly updated the get and add data so now you can calculate costs, it dynamically changes for electric and fuel types etc.