Updated the favicon and gave the site a name
Updated the favicon to be an electrical based icon instead of the vercel default and also fixed some bugs with the add data page's pagination. Re-added the "good morning" support to the user page as this was accidentally removed at some point. Added a space between the "good morning" and the username using padding.