Various updates
Updated the about page to include a little about myself, Added the electric page with more research and answers to common questions about electric cars and whether they should be purchased or not. Updated navbar and footer accordingly and expanded upon the HVO page substantially. Middleware updated to allow access without logging in